+++ actingweb/db/dynamodb/subscription_suspension.py
@@ -10,6 +10,7 @@ import os
 from datetime import UTC, datetime
 
 from pynamodb.attributes import UnicodeAttribute, UTCDateTimeAttribute
+from pynamodb.constants import PAY_PER_REQUEST_BILLING_MODE
 from pynamodb.exceptions import DoesNotExist
 from pynamodb.models import Model
 
@@ -25,8 +26,7 @@ class SubscriptionSuspension(Model):
         table_name = (
             os.getenv("AWS_DB_PREFIX", "demo_actingweb") + "_subscription_suspensions"
         )
-        read_capacity_units = 2
-        write_capacity_units = 1
+        billing_mode = PAY_PER_REQUEST_BILLING_MODE
         region = os.getenv("AWS_DEFAULT_REGION", "us-west-1")
         host = os.getenv("AWS_DB_HOST", None)